refactor: extract bbPress editor adapter (#76)
