+++ src/editor/index.tsx
@@ -22,6 +22,7 @@ import { createRegistry, RegistryProvider, useDispatch, useRegistry } from '@wor
  * Internal dependencies
  */
 import BuddyPress from './buddypress';
+import { createBbPressAdapter } from './bbpress-adapter';
 import ContentBridge from './content-bridge';
 import DetachedSidebar from './detached-sidebar';
 import EmbeddedEditorShell, { type ResolvedChromeConfig, type ResolvedToolbarConfig } from './embedded-editor-shell';
@@ -995,34 +996,29 @@ function createEditorContainer( container, textarea, settings ) {
 	const serviceContext = createServiceContext( settings, textarea, container );
 	const scopedApiFetch = createScopedApiFetch( services );
 
-	const bbpress = settings?.bbpress || {};
-	const bbpressIsTopicEdit = Boolean( bbpress?.isTopicEdit );
-	const bbpressIsReplyEdit = Boolean( bbpress?.isReplyEdit );
-	const bbpressTopicId = bbpress?.topicId ? Number( bbpress.topicId ) : 0;
-	const bbpressDraftEndpoint = bbpress?.draftEndpoint || null;
-	const bbpressMediaEndpoint = bbpress?.mediaEndpoint || null;
-	const blocksEverywhereMediaEndpoint = settings?.blocksEverywhere?.mediaUploadEndpoint || null;
-	const configuredMediaUploadEndpoint = bbpressMediaEndpoint || blocksEverywhereMediaEndpoint || null;
 	const hasUploadPermission = resolvePermission(
 		services,
 		'uploadMedia',
 		serviceContext,
 		settings?.editor?.hasUploadPermissions === true
 	);
-	const hasBbpressMediaUploadSupport = hasUploadPermission && Boolean( configuredMediaUploadEndpoint );
-
-	let currentForumId = bbpress?.forumId ? Number( bbpress.forumId ) : 0;
-	let autosaveTimer = null;
-	let lastSavedPayload = null;
-	let lastSerializedContent = '';
-	let isSubmitting = false;
-	let isContextSwitching = false;
 	let isUnmounted = false;
 	const editorKey = 0;
-	const draftRequestControllers = new Set< AbortController >();
 	const contentBridge = createContentBridgeController( textarea, settings );
 	let entityBridge = null;
 	let hasEditorFocus = false;
+	const bbpressAdapter =
+		settings?.editorType === 'bbpress'
+			? createBbPressAdapter( {
+					container,
+					settings,
+					textarea,
+					services,
+					serviceContext,
+					scopedApiFetch,
+					notifyService,
+			  } )
+			: null;
 
 	const emitLifecycle = ( name, detail = {} ) => {
 		dispatchLifecycleEvent( name, { container, detail, settings, textarea } );
@@ -1097,461 +1093,6 @@ function createEditorContainer( container, textarea, settings ) {
 		cleanupCallbacks.push( () => form.removeEventListener( 'submit', onSubmit ) );
 	}
 
-	const configuredNonce = settings?.restNonce || window?.wpApiSettings?.nonce || null;
-	const restHeaders = configuredNonce ? { 'X-WP-Nonce': configuredNonce } : {};
-
-	const requestDraft = async ( method, payload ) => {
-		const controller = new AbortController();
-		draftRequestControllers.add( controller );
-
-		const restRoot = settings?.restUrl || window?.wpApiSettings?.root || null;
-		if ( ! bbpressDraftEndpoint && ! restRoot ) {
-			throw new Error( 'Draft endpoint not configured.' );
-		}
-
-		const url = bbpressDraftEndpoint ? new URL( bbpressDraftEndpoint ) : new URL( restRoot );
-		if ( ( method === 'DELETE' || method === 'GET' ) && payload && typeof payload === 'object' ) {
-			Object.keys( payload ).forEach( ( key ) => {
-				if ( payload[ key ] === undefined || payload[ key ] === null ) {
-					return;
-				}
-				url.searchParams.set( key, String( payload[ key ] ) );
-			} );
-		}
-
-		try {
-			if ( services?.apiFetch || services?.apiFetchMiddleware || services?.apiFetchMiddlewares ) {
-				return scopedApiFetch( {
-					url: url.toString(),
-					method,
-					signal: controller.signal,
-					headers: {
-						...restHeaders,
-						'Content-Type': 'application/json',
-					},
-					body:
-						method === 'DELETE' || method === 'GET'
-							? undefined
-							: payload
-							? JSON.stringify( payload )
-							: undefined,
-				} );
-			}
-
-			const response = await window.fetch( url.toString(), {
-				method,
-				credentials: 'same-origin',
-				signal: controller.signal,
-				headers: {
-					...restHeaders,
-					'Content-Type': 'application/json',
-				},
-				body:
-					method === 'DELETE' || method === 'GET'
-						? undefined
-						: payload
-						? JSON.stringify( payload )
-						: undefined,
-			} );
-
-			if ( ! response.ok ) {
-				throw new Error( 'Draft request failed.' );
-			}
-
-			return response.json();
-		} finally {
-			draftRequestControllers.delete( controller );
-		}
-	};
-
-	const isTopicDraft = () => {
-		if ( settings?.editorType !== 'bbpress' ) {
-			return false;
-		}
-
-		if ( bbpressIsTopicEdit ) {
-			return false;
-		}
-
-		return Boolean( textarea?.name === 'bbp_topic_content' || document.getElementById( 'bbp_topic_title' ) );
-	};
-
-	const isReplyDraft = () => {
-		if ( settings?.editorType !== 'bbpress' ) {
-			return false;
-		}
-
-		if ( bbpressIsReplyEdit ) {
-			return false;
-		}
-
-		return Boolean( textarea?.name === 'bbp_reply_content' ) && bbpressTopicId > 0;
-	};
-
-	const getTopicTitle = () => {
-		const el = document.getElementById( 'bbp_topic_title' );
-		return el && 'value' in el ? String( el.value || '' ) : '';
-	};
-
-	const getForumIdFromDom = () => {
-		const el = document.getElementById( 'bbp_forum_id' );
-		if ( ! el || ! ( 'value' in el ) ) {
-			return 0;
-		}
-
-		const numberValue = Number( el.value );
-		return Number.isFinite( numberValue ) && numberValue >= 0 ? numberValue : 0;
-	};
-
-	const getReplyToFromDom = () => {
-		const replyToField = textarea?.closest?.( 'form' )?.querySelector?.( 'input[name="bbp_reply_to"]' );
-		if ( ! replyToField || ! ( 'value' in replyToField ) ) {
-			return 0;
-		}
-
-		const numberValue = Number( replyToField.value );
-		return Number.isFinite( numberValue ) && numberValue >= 0 ? numberValue : 0;
-	};
-
-	const isEffectivelyEmptyBlockContent = ( value ) => {
-		const content = String( value || '' ).trim();
-		if ( ! content ) {
-			return true;
-		}
-
-		const normalized = content
-			.replace( /<!--\s+wp:paragraph\s+-->/g, '' )
-			.replace( /<!--\s+\/wp:paragraph\s+-->/g, '' )
-			.replace( /<p>(?:\s|&nbsp;|&#160;|<br\s*\/?>)*<\/p>/gi, '' )
-			.replace( /\s+/g, '' );
-
-		return normalized === '';
-	};
-
-	const buildDraftPayload = ( contentOverride = null, forumIdOverride = null ) => {
-		const content = typeof contentOverride === 'string' ? contentOverride : textarea?.value || '';
-
-		if ( isReplyDraft() ) {
-			return {
-				type: 'reply',
-				topic_id: bbpressTopicId,
-				reply_to: getReplyToFromDom(),
-				content,
-			};
-		}
-
-		if ( isTopicDraft() ) {
-			const resolvedForumId = forumIdOverride !== null ? Number( forumIdOverride ) : currentForumId;
-			return {
-				type: 'topic',
-				forum_id: resolvedForumId,
-				title: getTopicTitle(),
-				content,
-			};
-		}
-
-		return null;
-	};
-	const hasInjectedAutosaveService = Object.prototype.hasOwnProperty.call( services, 'autosave' );
-	const injectedAutosave = services?.autosave;
-	const getAutosaveDelay = () => {
-		if ( injectedAutosave && typeof injectedAutosave === 'object' && typeof injectedAutosave.delay === 'number' ) {
-			return injectedAutosave.delay;
-		}
-
-		return 800;
-	};
-	const saveWithInjectedAutosave = async ( payload ) => {
-		if ( ! injectedAutosave ) {
-			return;
-		}
-
-		if ( typeof injectedAutosave === 'function' ) {
-			await injectedAutosave( payload, serviceContext );
-			return;
-		}
-
-		await injectedAutosave.save?.( payload, serviceContext );
-	};
-
-	const shouldAutorestoreDraft = () => {
-		if ( ! textarea ) {
-			return false;
-		}
-
-		const hasContent = ! isEffectivelyEmptyBlockContent( textarea.value || '' );
-		if ( hasContent ) {
-			return false;
-		}
-
-		if ( isTopicDraft() ) {
-			const titleInput = document.getElementById( 'bbp_topic_title' );
-			const titleValue = titleInput && 'value' in titleInput ? String( titleInput.value || '' ).trim() : '';
-			return titleValue === '';
-		}
-
-		return isReplyDraft();
-	};
-
-	const restoreDraftIfNeeded = async () => {
-		if ( ! shouldAutorestoreDraft() ) {
-			return;
-		}
-
-		try {
-			if ( isTopicDraft() ) {
-				const forumIdFromDom = getForumIdFromDom();
-				const response = await requestDraft( 'GET', {
-					type: 'topic',
-					forum_id: forumIdFromDom,
-					prefer_unassigned: true,
-				} );
-				const draft = response?.draft;
-				if ( ! draft ) {
-					return;
-				}
-
-				const titleInput = document.getElementById( 'bbp_topic_title' );
-				if ( titleInput && 'value' in titleInput && String( titleInput.value || '' ).trim() === '' ) {
-					titleInput.value = String( draft?.title || '' );
-				}
-
-				if ( isEffectivelyEmptyBlockContent( textarea.value || '' ) ) {
-					textarea.value = String( draft?.content || '' );
-					lastSerializedContent = textarea.value;
-				}
-				return;
-			}
-
-			if ( isReplyDraft() && bbpressTopicId ) {
-				const response = await requestDraft( 'GET', {
-					type: 'reply',
-					topic_id: bbpressTopicId,
-					reply_to: getReplyToFromDom(),
-				} );
-				const draft = response?.draft;
-				if ( ! draft ) {
-					return;
-				}
-
-				if ( isEffectivelyEmptyBlockContent( textarea.value || '' ) ) {
-					textarea.value = String( draft?.content || '' );
-					lastSerializedContent = textarea.value;
-				}
-			}
-		} catch ( error ) {
-			if ( error?.name === 'AbortError' ) {
-				return;
-			}
-			// eslint-disable-next-line no-console
-			console.error( 'Blocks Everywhere: failed to restore bbPress draft', error );
-		}
-	};
-
-	const scheduleAutosaveFromContent = ( content, forumIdOverride = null ) => {
-		if ( isSubmitting || isContextSwitching ) {
-			return;
-		}
-
-		lastSerializedContent = typeof content === 'string' ? content : '';
-		if ( hasInjectedAutosaveService ) {
-			if ( injectedAutosave === null ) {
-				return;
-			}
-
-			const draft = buildDraftPayload( lastSerializedContent, forumIdOverride );
-			const payload = draft || {
-				content: lastSerializedContent,
-				editorType: settings?.editorType || '',
-				textareaName: textarea?.name || '',
-			};
-			const payloadString = JSON.stringify( payload );
-
-			if ( payloadString === lastSavedPayload ) {
-				return;
-			}
-
-			if ( autosaveTimer ) {
-				clearTimeout( autosaveTimer );
-			}
-
-			autosaveTimer = setTimeout( async () => {
-				if ( isSubmitting || isContextSwitching ) {
-					return;
-				}
-
-				try {
-					await saveWithInjectedAutosave( payload );
-					lastSavedPayload = payloadString;
-				} catch ( error ) {
-					if ( error?.name === 'AbortError' ) {
-						return;
-					}
-
-					notifyService( services, 'error', 'Autosave failed.', serviceContext, error );
-					// eslint-disable-next-line no-console
-					console.error( 'Blocks Everywhere: injected autosave failed', error );
-				}
-			}, getAutosaveDelay() );
-			return;
-		}
-
-		const draft = buildDraftPayload( lastSerializedContent, forumIdOverride );
-		if ( ! draft ) {
-			return;
-		}
-
-		const hasAnyContent =
-			Boolean( String( draft?.content || '' ).trim() ) || Boolean( String( draft?.title || '' ).trim() );
-		if ( ! hasAnyContent ) {
-			return;
-		}
-
-		const payloadString = JSON.stringify( draft );
-		if ( payloadString === lastSavedPayload ) {
-			return;
-		}
-
-		if ( autosaveTimer ) {
-			clearTimeout( autosaveTimer );
-		}
-
-		autosaveTimer = setTimeout( async () => {
-			if ( isSubmitting || isContextSwitching ) {
-				return;
-			}
-
-			try {
-				await requestDraft( 'POST', draft );
-				lastSavedPayload = payloadString;
-			} catch ( error ) {
-				if ( error?.name === 'AbortError' ) {
-					return;
-				}
-				// eslint-disable-next-line no-console
-				console.error( 'Blocks Everywhere: bbPress draft autosave failed', error );
-			}
-		}, 800 );
-	};
-
-	const scheduleAutosave = ( serializedContent ) => {
-		scheduleAutosaveFromContent( typeof serializedContent === 'string' ? serializedContent : '' );
-	};
-
-	const maybeInstallForumMoveHandler = () => {
-		if ( ! isTopicDraft() ) {
-			return;
-		}
-
-		const forumSelect = document.getElementById( 'bbp_forum_id' );
-		if ( ! forumSelect || forumSelect.__blocksEverywhereDraftMoveInstalled ) {
-			return;
-		}
-
-		forumSelect.__blocksEverywhereDraftMoveInstalled = true;
-		currentForumId = getForumIdFromDom();
-
-		if ( isTopicDraft() ) {
-			const titleInput = document.getElementById( 'bbp_topic_title' );
-			if ( titleInput && ! titleInput.__blocksEverywhereDraftTitleInstalled ) {
-				titleInput.__blocksEverywhereDraftTitleInstalled = true;
-				const handler = () => {
-					scheduleAutosaveFromContent( lastSerializedContent || textarea?.value || '' );
-				};
-
-				titleInput.addEventListener( 'input', handler );
-				cleanupCallbacks.push( () => {
-					titleInput.removeEventListener( 'input', handler );
-					delete titleInput.__blocksEverywhereDraftTitleInstalled;
-				} );
-			}
-		}
-
-		const handler = async () => {
-			if ( isSubmitting ) {
-				return;
-			}
-
-			const nextForumId = getForumIdFromDom();
-			const previousForumId = currentForumId;
-			currentForumId = nextForumId;
-
-			if ( previousForumId !== 0 || nextForumId <= 0 ) {
-				return;
-			}
-
-			const draft = buildDraftPayload( lastSerializedContent || null, nextForumId );
-			if ( ! draft ) {
-				return;
-			}
-
-			const hasAnyContent =
-				Boolean( String( draft?.content || '' ).trim() ) || Boolean( String( draft?.title || '' ).trim() );
-			if ( ! hasAnyContent ) {
-				return;
-			}
-
-			try {
-				await requestDraft( 'POST', draft );
-				await requestDraft( 'DELETE', { type: 'topic', forum_id: 0 } );
-				lastSavedPayload = JSON.stringify( draft );
-			} catch ( error ) {
-				if ( error?.name === 'AbortError' ) {
-					return;
-				}
-				// eslint-disable-next-line no-console
-				console.error( 'Blocks Everywhere: failed to move forum draft', error );
-			}
-		};
-
-		forumSelect.addEventListener( 'change', handler );
-		cleanupCallbacks.push( () => {
-			forumSelect.removeEventListener( 'change', handler );
-			delete forumSelect.__blocksEverywhereDraftMoveInstalled;
-		} );
-	};
-
-	const uploadViaConfiguredMediaEndpoint = async ( file, topicId = 0 ) => {
-		const formData = new FormData();
-		formData.append( 'file', file );
-		formData.append( 'context', 'content_embed' );
-		if ( topicId ) {
-			formData.append( 'target_id', String( topicId ) );
-		}
-
-		const uploadNonce = settings?.restNonce || window?.wpApiSettings?.nonce || null;
-		const headers = uploadNonce ? { 'X-WP-Nonce': uploadNonce } : undefined;
-
-		if ( ! configuredMediaUploadEndpoint ) {
-			throw new Error( 'Media endpoint not configured.' );
-		}
-
-		const response = await window.fetch(
-			new URL( configuredMediaUploadEndpoint, window.location.origin ).toString(),
-			{
-				method: 'POST',
-				credentials: 'same-origin',
-				headers,
-				body: formData,
-			}
-		);
-
-		if ( ! response.ok ) {
-			let errorMessage = 'Upload failed.';
-			try {
-				const payload = await response.json();
-				if ( payload?.message ) {
-					errorMessage = payload.message;
-				}
-			} catch ( error ) {
-				// ignore
-			}
-			throw new Error( errorMessage );
-		}
-
-		return response.json();
-	};
-
 	const renderEditor = () => {
 		// Opt-in postEntity wiring: when the consumer declares this BE mount is
 		// backed by a canonical WP post, wrap the editor in <EditorProvider> so
@@ -1589,14 +1130,14 @@ function createEditorContainer( container, textarea, settings ) {
 							const serialized = contentBridge.save( newBlocks );
 							entityBridge.saveEdits( newBlocks, serialized, 'input' );
 							emitContentHook( 'input', newBlocks, serialized );
-							scheduleAutosave( serialized );
+							bbpressAdapter?.scheduleAutosave( serialized );
 						} }
 						onChange={ ( newBlocks ) => {
 							settings?.blocksEverywhere?.__experimentalOnChange?.( newBlocks );
 							const serialized = contentBridge.save( newBlocks );
 							entityBridge.saveEdits( newBlocks, serialized, 'change' );
 							emitContentHook( 'change', newBlocks, serialized );
-							scheduleAutosave( serialized );
+							bbpressAdapter?.scheduleAutosave( serialized );
 						} }
 						onSelection={ ( selection ) =>
 							settings?.blocksEverywhere?.__experimentalOnSelection?.( selection )
@@ -1634,140 +1175,7 @@ function createEditorContainer( container, textarea, settings ) {
 		);
 	};
 
-	const maybeInstallSubmitHandler = () => {
-		if ( settings?.editorType !== 'bbpress' ) {
-			return;
-		}
-
-		if ( ! container || container.__blocksEverywhereDraftSubmitInstalled ) {
-			return;
-		}
-
-		const form = container.closest( 'form' );
-		if ( ! form ) {
-			return;
-		}
-
-		container.__blocksEverywhereDraftSubmitInstalled = true;
-		const handler = ( event ) => {
-			if ( event.submitter && event.submitter.closest( '.blocks-everywhere-editor' ) ) {
-				return;
-			}
-
-			isSubmitting = true;
-			if ( autosaveTimer ) {
-				clearTimeout( autosaveTimer );
-			}
-
-			draftRequestControllers.forEach( ( controller ) => controller.abort() );
-		};
-
-		form.addEventListener( 'submit', handler );
-		cleanupCallbacks.push( () => {
-			form.removeEventListener( 'submit', handler );
-			delete container.__blocksEverywhereDraftSubmitInstalled;
-		} );
-	};
-
-	const maybeInstallReplyDraftContextHandler = () => {
-		if ( ! isReplyDraft() ) {
-			return;
-		}
-
-		if ( ! container || container.__blocksEverywhereReplyDraftContextInstalled ) {
-			return;
-		}
-
-		container.__blocksEverywhereReplyDraftContextInstalled = true;
-
-		const handler = async ( event ) => {
-			if ( ! event?.detail || event.detail.type !== 'reply' ) {
-				return;
-			}
-
-			const topicId = event.detail.topicId ? Number( event.detail.topicId ) : 0;
-			if ( ! topicId || topicId !== bbpressTopicId ) {
-				return;
-			}
-
-			if ( isSubmitting ) {
-				return;
-			}
-
-			const previousReplyTo = event.detail.previousReplyTo ? Number( event.detail.previousReplyTo ) : 0;
-			const nextReplyTo = event.detail.nextReplyTo ? Number( event.detail.nextReplyTo ) : 0;
-
-			if ( previousReplyTo === nextReplyTo ) {
-				return;
-			}
-
-			isContextSwitching = true;
-			if ( autosaveTimer ) {
-				clearTimeout( autosaveTimer );
-			}
-
-			draftRequestControllers.forEach( ( controller ) => controller.abort() );
-
-			try {
-				const outgoingContent = String( lastSerializedContent || textarea?.value || '' );
-				if ( outgoingContent.trim() ) {
-					const outgoingPayload = {
-						type: 'reply',
-						topic_id: bbpressTopicId,
-						reply_to: previousReplyTo,
-						content: outgoingContent,
-					};
-					await requestDraft( 'POST', outgoingPayload );
-					lastSavedPayload = JSON.stringify( outgoingPayload );
-				}
-
-				lastSerializedContent = '';
-				lastSavedPayload = null;
-
-				const incoming = await requestDraft( 'GET', {
-					type: 'reply',
-					topic_id: bbpressTopicId,
-					reply_to: nextReplyTo,
-				} );
-
-				const incomingDraft = incoming?.draft;
-				const incomingContent =
-					incomingDraft && String( incomingDraft?.content || '' ).trim()
-						? String( incomingDraft.content )
-						: '';
-
-				// Use the ContentBridge API to hot-swap content without remounting.
-				const contentApi = textarea?.__blocksEverywhereContentApi;
-				if ( contentApi ) {
-					contentApi.replaceContent( incomingContent );
-				}
-
-				// Keep textarea in sync for onSaveContent and autosave tracking.
-				textarea.value = incomingContent;
-				lastSerializedContent = incomingContent;
-			} catch ( error ) {
-				if ( error?.name === 'AbortError' ) {
-					return;
-				}
-				// eslint-disable-next-line no-console
-				console.error( 'Blocks Everywhere: failed to switch reply draft context', error );
-			} finally {
-				isContextSwitching = false;
-			}
-		};
-
-		document.addEventListener( 'blocksEverywhere:bbpressDraftContextChange', handler );
-		cleanupCallbacks.push( () => {
-			document.removeEventListener( 'blocksEverywhere:bbpressDraftContextChange', handler );
-			delete container.__blocksEverywhereReplyDraftContextInstalled;
-		} );
-	};
-
-	if ( settings?.editorType === 'bbpress' ) {
-		maybeInstallForumMoveHandler();
-		maybeInstallSubmitHandler();
-		maybeInstallReplyDraftContextHandler();
-	}
+	bbpressAdapter?.installHandlers();
 
 	if ( services?.fetchLinkSuggestions !== undefined ) {
 		settings.editor.__experimentalFetchLinkSuggestions = services.fetchLinkSuggestions || undefined;
@@ -1780,16 +1188,15 @@ function createEditorContainer( container, textarea, settings ) {
 			addFilter( 'editor.MediaUpload', 'blocks-everywhere/media-upload', () => MediaUpload );
 		}
 	} else if ( settings?.editorType === 'bbpress' ) {
-		if ( ! hasBbpressMediaUploadSupport ) {
+		if ( ! hasUploadPermission || ! bbpressAdapter?.mediaEndpoint ) {
 			settings.editor.mediaUpload = null;
 		} else {
 			settings.editor.mediaUpload = ( { filesList, onFileChange, onError } ) => {
 				const files = Array.from( filesList );
-				const topicId = bbpressTopicId;
 
 				Promise.all(
 					files.map( async ( file ) => {
-						const result = await uploadViaConfiguredMediaEndpoint( file, topicId );
+						const result = await bbpressAdapter.uploadMedia( file );
 						const attachment = result?.attachment;
 						if ( attachment ) {
 							return attachment;
@@ -1822,7 +1229,7 @@ function createEditorContainer( container, textarea, settings ) {
 	void ( async () => {
 		try {
 			emitLifecycle( 'before-load', { instance } );
-			await restoreDraftIfNeeded();
+			await bbpressAdapter?.restoreDraftIfNeeded();
 			if ( isUnmounted ) {
 				return;
 			}
@@ -1842,15 +1249,7 @@ function createEditorContainer( container, textarea, settings ) {
 		emitLifecycle( 'before-unmount', { instance } );
 		isUnmounted = true;
 
-		if ( autosaveTimer ) {
-			clearTimeout( autosaveTimer );
-		}
-
-		if ( injectedAutosave && typeof injectedAutosave === 'object' ) {
-			injectedAutosave.cancel?.( serviceContext );
-		}
-
-		draftRequestControllers.forEach( ( controller ) => controller.abort() );
+		bbpressAdapter?.cleanup();
 		container?.removeEventListener?.( 'focusin', onFocusIn );
 		container?.removeEventListener?.( 'focusout', onFocusOut );
 		cleanupCallbacks.forEach( ( cleanup ) => cleanup() );
